+++ operator/internal/component/registry_test.go
